Fix viser-ng exports
